+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.31.3] - 2025-09-22
+
+Add variant `Cleanup::KeepForDays(usize)` to have log files being deleted after a given number of days.
+Kudos goes to [Tunglies](https://github.com/Tunglies).
+
 ## [0.31.2] - 2025-06-27
 
 Add syslog-call based `SyslogWriter` connection